+++ src/pages/market/MarketListItem.jsx
@@ -10,7 +10,7 @@ function MarketListItem({ item }) {
     <>
       <li>
         <Link className="card" to={`/market/detail/${item._id}`}>
-          <img className="card-cover" src={`${import.meta.env.VITE_API_SERVER}/files/${import.meta.env.VITE_CLIENT_ID}/${item?.mainImages?.fileName}` || null} alt={`${item.name} 상품 사진`} />
+          <img className="card-cover" src={`${import.meta.env.VITE_API_SERVER}/files/${import.meta.env.VITE_CLIENT_ID}/${item?.mainImages[0].fileName}` || null} alt={`${item.name} 상품 사진`} />
           <div className="card-contents">
             <span className="card-brand">{item.extra?.brand}</span>
             <h4 className="card-title mini">{item.name}</h4>
